Remove unnecessary casts because ctors now easier to use

+++ src/test/java/nom/bdezonia/zorbage/algorithm/TestCopyIf.java
@@ -51,7 +51,7 @@ public class TestCopyIf {
 				new short[] {1,2,3,4,5,6,7,8,9});
 		IndexedDataSource<?, SignedInt16Member> b = ArrayStorage.allocateShorts(
 				new short[] {0,0,0,0,0,0,0,0,0});
-		Condition<SignedInt16Member> cond = new GreaterThanConstant<SignedInt16Group,SignedInt16Member>(G.INT16, new SignedInt16Member((short)4));
+		Condition<SignedInt16Member> cond = new GreaterThanConstant<SignedInt16Group,SignedInt16Member>(G.INT16, new SignedInt16Member(4));
 		CopyIf.compute(G.INT16, cond, a, b);
 		SignedInt16Member value = G.INT16.construct();
 		b.get(0, value);
